+++ package.json
@@ -19,6 +19,7 @@
   },
   "devDependencies": {
     "@alpinejs/collapse": "^3.5.1",
+    "@types/bn.js": "^5.1.0",
     "@types/chrome": "^0.0.188",
     "@types/debug": "^4.1.7",
     "@types/jest": "^28.1.1",
@@ -29,20 +30,24 @@
     "cross-env": "^7.0.0",
     "debug": "^4.3.3",
     "husky": "^8.0.1",
+    "jest": "^28.1.1",
+    "jest-chrome": "^0.7.2",
+    "jest-fetch-mock": "^3.0.3",
+    "jest-ts-auto-mock": "^2.1.0",
     "lint-staged": "^13.0.1",
     "loader-utils": "^2.0.0",
     "postcss-cli": "^7.1.0",
     "postcss-loader": "^3.0.0",
     "prettier": "^2.6.2",
     "tailwindcss": "^1.1.4",
+    "ts-auto-mock": "^3.6.2",
+    "ts-jest": "^28.0.5",
     "ts-loader": "~8.2.0",
-    "typescript": "^4.5.5",
+    "ts-node": "^10.8.0",
+    "ttypescript": "^1.5.13",
+    "typescript": "^4.7.3",
     "webpack": "4.41.5",
     "webpack-cli": "3.3.10",
-    "webpack-extension-reloader": "^1.1.4",
-    "jest": "^28.1.0",
-    "jest-chrome": "^0.7.2",
-    "ts-jest": "^28.0.3",
-    "ts-node": "^10.8.0"
+    "webpack-extension-reloader": "^1.1.4"
   }
 }
