+++ tsconfig.json
@@ -14,6 +14,12 @@
       "types": [
         "jest",
         "chrome",
+      ],
+      "plugins": [
+        {
+          "transform": "ts-auto-mock/transformer",
+          "cacheBetweenTests": false
+        }
       ]
   },
   "include": [
